pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
backend-path = ["."]

[project]
name = "WeatherCaster"
version = "1.0.0"
description = "An AI-powered command-line chatbot that provides weather forecasts using a local LLM and OpenWeatherMap API."
readme = "README.md"
authors = [
    { name = "Yücel Muhammed", email = "Muhammed.Yuecel@study-iu.org" }
]
requires-python = ">=3.12" 

dependencies = [
    "pydantic-ai>=0.2.4",
    "python-dotenv>=0.21.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.28.0",
    "gradio>=5.30.0",
]

classifiers = [
    "Development Status :: 4 - Beta", # Or "3 - Alpha", "5 - Production/Stable"
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Communications :: Chat",
    "Operating System :: OS Independent",
]

[project.urls]
"Homepage" = "https://github.com/MYCL94/WeatherCaster"
"Repository" = "https://github.com/MYCL94/WeatherCaster"


[project.scripts]
weathercaster-cli = "weather_caster.cli:run_cli_sync_wrapper"
weathercaster-gui = "weather_caster.gradio_ui:run_gradio_ui_sync_wrapper"
